Various fixes for 3.69 aws provider

+++ scripts/get-eks-cluster-nodegroups.sh
@@ -61,21 +61,21 @@ if [ "$kcount" -gt "0" ]; then
                            
                             #echo "cname=$cname ngname=$ngnam ocname=$ocname"
                             fn=`printf "%s__%s.tf" $ttft $cname`
+                            echo "$ttft $cname import"
+                            if [ -f "$fn" ]; then continue; fi
 
-                            printf "resource \"%s\" \"%s\" {" $ttft $cname > $ttft.$cname.tf
-                            printf "}" >> $ttft.$cname.tf
+                            printf "resource \"%s\" \"%s\" {" $ttft $cname > $fn
+                            printf "}" >> $fn
                             #echo "pre-import"
                             #ls -l
                             echo "Importing ....."
                             terraform import $ttft.$cname $ocname | grep Import
                             terraform state show $ttft.$cname > t2.txt
-                            rm $ttft.$cname.tf
+                            rm $fn
                             cat t2.txt | perl -pe 's/\x1b.*?[mGKH]//g' > t1.txt
                             mv="version"
                             file="t1.txt"
                             
-                            fn=`printf "%s__%s.tf" $ttft $cname`
-                
                             echo $aws2tfmess > $fn
                             iscust=0
                             allowid=0
@@ -137,6 +137,21 @@ if [ "$kcount" -gt "0" ]; then
                                         #if [ "$iscust" == "1" ]; then
                                             skip=1;
                                         #fi
+                                    fi
+
+                                    if [[ ${tt1} == "max_unavailable_percentage" ]];then 
+                                        tt2=`echo $tt2 | tr -d ' '`
+                                        if [ "$tt2" == "0" ]; then
+                                            skip=1;
+                                        fi
+                                    fi
+
+
+                                    if [[ ${tt1} == "max_unavailable" ]];then 
+                                        tt2=`echo $tt2 | tr -d ' '`
+                                        if [ "$tt2" == "0" ]; then
+                                            skip=1;
+                                        fi
                                     fi
 
                                     if [[ ${tt1} == "resources" ]];then
